refactor: Move diskalloc -> writernode/diskalloc
